scale: Unpack each input row in a separate step
This avoids repeatedly performing the unpack operation when pixels are
being reused. It's also more SIMD-friendly, since the inner loops are
simpler and we can align the unpacked data as we please.

It can be slightly slower for bigger inputs (e.g. 4096+ pixels wide), since
the unpacked temporary data puts more pressure on the L1 cache. Oh well.